feat!: improve the package structure
