+++ .github/workflows/cd.yml
@@ -28,5 +28,6 @@ jobs:
         with:
           version: latest
           args: release --rm-dist
+          workdir: cmd/vulcain
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
