+++ .gitignore
@@ -1,4 +1,4 @@
-/vulcain
+/cmd/vulcain/vulcain
 /spec/vulcain.html
 /spec/vulcain.txt
 /spec/vulcain.xml
