+++ cmd/vulcain/.goreleaser.yml
