+++ go.mod
@@ -1,9 +1,10 @@
 module github.com/dunglas/vulcain
 
-go 1.13
+go 1.15
 
 require (
 	github.com/dunglas/httpsfv v0.0.0-20200829121517-5350feba44b9
+	github.com/francoispqt/gojay v1.2.13
 	github.com/getkin/kin-openapi v0.22.0
 	github.com/gofrs/uuid v3.3.0+incompatible
 	github.com/golang/protobuf v1.4.2 // indirect
@@ -15,6 +16,7 @@ require (
 	github.com/stretchr/testify v1.5.1
 	github.com/tidwall/gjson v1.6.1
 	github.com/tidwall/sjson v1.1.1
+	gocloud.dev v0.20.0
 	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a
 	golang.org/x/net v0.0.0-20200904194848-62affa334b73
 	golang.org/x/sys v0.0.0-20200909081042-eff7692f9009 // indirect
