+++ openapi.go
@@ -1,4 +1,4 @@
-package gateway
+package vulcain
 
 import (
 	"net/url"
